+++ src/collective/big/bang/interface.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from zope.lifecycleevent import IObjectModifiedEvent
+
+
+class IDarwinStartedEvent(IObjectModifiedEvent):
+    """
+    Site is created an admin password updated.
+    """